refactor(components): refactor overlay
