+++ packages/components/drawer/src/index.vue
@@ -59,7 +59,7 @@
 
 <script lang="ts">
 import { defineComponent, computed, ref } from 'vue'
-import { Overlay } from '@element-plus/components/overlay'
+import { ElOverlay } from '@element-plus/components/overlay'
 import {
   useDialog,
   dialogProps,
@@ -74,7 +74,7 @@ type DrawerDirection = 'ltr' | 'rtl' | 'ttb' | 'btt'
 export default defineComponent({
   name: 'ElDrawer',
   components: {
-    [Overlay.name]: Overlay,
+    ElOverlay,
   },
   directives: {
     TrapFocus,
