+++ packages/components/message-box/src/index.vue
@@ -128,7 +128,7 @@ import {
   usePreventGlobal,
 } from '@element-plus/hooks'
 import ElInput from '@element-plus/components/input'
-import { Overlay as ElOverlay } from '@element-plus/components/overlay'
+import { ElOverlay } from '@element-plus/components/overlay'
 import PopupManager from '@element-plus/utils/popup-manager'
 import { on, off } from '@element-plus/utils/dom'
 import { EVENT_CODE } from '@element-plus/utils/aria'
