+++ packages/components/overlay/src/overlay.ts
@@ -1,27 +1,38 @@
-<script lang="ts">
 import { createVNode, defineComponent, renderSlot, h } from 'vue'
 import { PatchFlags } from '@element-plus/utils/vnode'
 import { useSameTarget } from '@element-plus/hooks'
+import { buildProp } from '@element-plus/utils/props'
+
+import type { ExtractPropTypes, CSSProperties } from 'vue'
+
+export const overlayProps = {
+  mask: {
+    type: Boolean,
+    default: true,
+  },
+  customMaskEvent: {
+    type: Boolean,
+    default: false,
+  },
+  overlayClass: buildProp<string | string[] | Record<string, boolean>>({
+    type: [String, Array, Object],
+  }),
+  zIndex: buildProp<CSSProperties['zIndex']>({
+    type: [String, Number],
+  }),
+} as const
+export type OverlayProps = ExtractPropTypes<typeof overlayProps>
+
+export const overlayEmits = {
+  click: (evt: MouseEvent) => evt instanceof MouseEvent,
+}
+export type OverlayEmits = typeof overlayEmits
 
 export default defineComponent({
   name: 'ElOverlay',
-  props: {
-    mask: {
-      type: Boolean,
-      default: true,
-    },
-    customMaskEvent: {
-      type: Boolean,
-      default: false,
-    },
-    overlayClass: {
-      type: [String, Array, Object],
-    },
-    zIndex: {
-      type: Number,
-    },
-  },
-  emits: ['click'],
+
+  props: overlayProps,
+  emits: overlayEmits,
 
   setup(props, { slots, emit }) {
     const onMaskClick = (e: MouseEvent) => {
@@ -63,11 +74,10 @@ export default defineComponent({
                 right: '0px',
                 bottom: '0px',
                 left: '0px',
-              },
+              } as CSSProperties,
             },
             [renderSlot(slots, 'default')]
           )
     }
   },
 })
-</script>
